IMPULSE - 6 Read mp4 files. Add More Diagnostics.
The mp4 files require the OpenCV to be compiled with ffmpeg support. Sadly
OpenCV provides no clear way to actually check whether it has such support
at runtime. So we attempt to open the file and on failure we suggest to the
user that they ensure that OpenCV has ffmpeg support.

Fixes https://github.com/cwgreene/Impulse/issues/6